dns-config - new role
For now, this is geared towards enabling DNS in the libvirt VMs.
We like libvirt because GPU passthrough is easier. But it's missing the
handy one-liner to turn on DNS proxying that Virtualbox has.

To get around that, let's just implement a simple role to setup DNS
manually, without relying on a proxy. It might be useful in other
scenarios, too.