Add a simple Slurm + MPI Hello, World example
(Take 2! This is a re-do after I messed up PR #160.)

Getting the examples section started with a very simple MPI example,
demonstrated to work in a virtual Slurm DeepOps cluster.

While we will want to have some more impressive examples, e.g. Deep
Learning training and all that jazz, we should also have some simple and
clear examples that are easy for new users to build on and get running.
This example is about as simple as you can get while still being an MPI,
but also provides a Slurm usage example.

Next one I'm working on is an equivalent "hello, world" Flask service
running in k8s, but running into fun with the registry which is delaying
that. So might as well open a PR for this on its own. :)