Add a check for required linters to enable script
If you don't have one or more linters installed, you'll get annoying
errors when you try to commit. So let's check for the linters before we
enable the hook.

I went back and forth about including the install commands directly, but
decided to go for principle of least surprise and let the user take care
of installs if they want.