Purge + re-install galaxy roles in virtual cluster
I've had a couple cases where a galaxy dependency didn't correctly make
it into the `requirements.yml`, but I failed to notice because it was
cached locally in `galaxy-roles/`. This is especially problematic when
working in the virtual cluster, which is commonly used for test
workflows and CI.

This change just nukes the `galaxy-roles/` directory as part of
`cluster_up.sh`, and ensures that all dependencies are re-installed
fresh.

Cons to this change:

* This adds some time (~20-40 s) to cluster turnup, but htis is a small
fraction of the total time.
* This could cause `cluster_up.sh` to fail if a download fails. But
we're downloading a ton of data in the whole turnup process, so we're
already exposed to this risk, and this at least fails at the beginning
of the process.