Set up MetalLB by default in virtual cluster
There are a lot of options when it comes to setting up external access
to a k8s cluster, but a simple L2 load balancer is a good default
option. So let's set it up by default, and let the user make changes
later if they like.

This follows the general principle that a virtual cluster should be as
useful as possible when turned up with the default config.