Optionally clean up virtual config on cluster_up
Continuing to try to make the virtual cluster development experience a
little better... I've had some recent cases where I made changes to
`config.example` or to the k8s config generation process, and had to
manually whack the auto-copied config to get my changes to take place.

The virtual cluster should be in a clean state when we run
`cluster_up.sh`, so let's just explicitly clean up these directories so
we can start with fresh copies.

This is conditional with a `DEEPOPS_VIRT_CLEAN_CONFIG` variable that can
be used to prevent the cleanup if needed.