[slurm] Refactor prolog+epilog files
1. Use a single file under prolog.d/ and epilog.d/.

Slurm lists files in reverse alphabetic order (why?). run-parts on
Ubuntu can use the --reverse flag to mimic this behavior, but run-parts
on centos does not have this flag. To reduce confusion, we skip slurm's
reverse ordering by only giving it one file to work with. Then, we
arrange the rest of the prolog+epilog files to run in the normal order.

2. Use a single "-parts.d/" directory for prolog+epilog

In the previous setup, ALL of the "exclusive" tasks ran before ALL of the
"lastuser" tasks. To be more flexible, we now use a single directory.
Now, with a simple `ls` you can see the order in which tasks will be
run. To switch on/off the exclusive/lastuser tasks, we use keywords in
the filenames. On ubuntu, run-parts has a regex feature which we could
use to run only files with certain keywords. But, again, run-parts on
centos has fewer features. So, we simply ship a simple version of
run-parts with DeepOps.

3. Tweak the ordering of tasks

It seems to me this ordering will cause the most reliable shutdown. It
may need to be tweaked further.

* Kill docker containers
  * Afterwards, cleanup docker images
* Kill user processes
* Kill GPU processes
* Cleanup tmpfiles

4. Add option to purge old prolog+epilog files