Fix script usage of `apt` to use `apt-get`
`apt` is a user-facing tool which does not guarantee a stable command
line API. `apt-get` is the recommended tool for use in scripts.

This is a minor issue, but it bugs me every time I watch `apt` print
that warning, and it does provide a little future-proofing for the
scripts. :-)