Temporary work-around for Kubespray issue 6160
https://github.com/kubernetes-sigs/kubespray/issues/6160

Kubespray pins the version of `docker-ce`, but does not pin the version
of `docker-ce-cli` by default.

Unfortunately, Docker has published a version of `docker-ce-cli` (19.03.9)
which is no longer compatible with the pinned `docker-ce` version
(18.09.7). When the two are installed together, any docker cli commands
like `docker images` result in an error like:

```
Error response from daemon: client version 1.40 is too new. Maximum
supported API version is 1.39
```

This PR adds configuration for the Kubespray docker role to pin the
`docker-ce-cli` version as well as the `docker-ce` version.

We will need to track the Kubespray issue so that we can remove this
WAR when it's fixed upstream.