Ensure rpc-statd is enabled and running
Spack uses advisory locks when running on shared filesystems, which
requires rpc-statd to be running on the server. This isn't enabled by
default, at least on Ubuntu, so we need to ensure it's started.

Having it running is a good idea anyway, so we'll just add it to the
role by default.