Temporarily take Spack out of CI
Needs further debugging to understand why this is failing in CI, as it's
failing on an ansible-playbook command that should "just work". We
haven't even installed anything from Spack yet, so there must be some
earlier change mucking about with Ansible.

Let's take this out of the active CI path for now, and then I'll do some
additional debugging in a separate PR.