[nvidia-mig] Remove unnecessary steps
Enabling MIG mode requires a GPU reset, but nvidia-smi -mig 0/1 takes
care of it already, as long as there are no daemons or other processes
getting in the way. There are also two daemons that do handshake
explicitly with nvidia-smi for GPU reset, namely nvidia-fabricmanager
and nvidia-persistenced, and they do not need to be stopped.

It's also unnecessary to unload any nvidia modules, or stop the
nv_peer_memory service that only loads/unloads modules.

Fix all of that. And also make sure to only restart daemons that were
actually running when the playbook was executed, and stop enabling any
of them.