[nvidia-mig] Only reset services if MIG mode changes are needed
If all GPUs have the desired state already, skip shutting down and
restarting various system services.

Also explicitly fail if none of the GPUs support MIG and it was
requested to be enabled. Requesting MIG to be disabled in such case is
considered ok as GPUs with no MIG have it always disabled.