Ignore errors when copying slurm service files
When this playbook is run on a cluster where slurm is already installed,
it will skip the slurm build. This is good, but if we have cleaned up
the build in the mean time then the tasks for copying the slurm service
files (which is separate from the build) will fail, killing the whole
playbook run.

This change ignores errors on the step where we copy the service files,
so that the playbook can run correctly in the case of an
already-installed cluster.

It should be safe to ignore errors for copying these files, because
if they're absent then the playbook will fail later when we try to start
the services.