Fix for all GPUs visible when ssh on compute node
As mentioned in https://slurm.schedmd.com/pam_slurm_adopt.html, "The pam_systemd module will conflict with pam_slurm_adopt, so you need to disable it in all files that are included in sshd or system-auth (e.g. password-auth, common-session, etc.)." 

Without disabling it, when a user started a job, they could ssh to the compute node and while there see all GPUs that are installed on the system even though their job did not specifically request for them.