Add a GPU job test for Slurm
Right now we don't have a Jenkins test that executes a GPU job under the
Slurm scheduler. I think we validate GPU functionality pretty well under
Kubernetes, but it would be good to check it under Slurm as well.

This test uses the CUDA toolkit in nvhpc to build the `deviceQuery`
sample and run it on the GPU node.