Fix disabling NVIDIA driver install when desired
The Slurm cluster playbook provides the var slurm_cluster_install_nvidia_driver to disable installs of the NVIDIA driver.
However, the Slurm cluster playbook also sources the CUDA playbook, which installs the driver itself.
And the CUDA playbook does not respect the flag in the Slurm cluster playbook.

Aren't dependencies fun?

This commit adds a flag to the CUDA playbook to allow disabling the driver install, and plumbs it through so the user can set the flag in the Slurm cluster playbook.

To test, set slurm_cluster_install_nvidia_driver to false and set up a Slurm cluster, and the driver should not be installed.