Permit SFTP in default SSHD config
Our SSH hardening role disables the use of SFTP by default. This causes
problems with Ansible, which expects to be able to use SFTP to transfer
large files. In particular, this has caused problems for the DGX
firmware upgrade role, which will fail to work correctly if the SSH
hardening role has run.

This commit also makes it easier to override the SSH configuration
by moving hard-coded parameters from the user playbook to the DeepOps
configuration.