Add basic role for Squid HTTP proxy
This commit adds a basic role to set up an HTTP proxy using
[Squid](http://www.squid-cache.org/).

This role is primarily intended to make it easier to test deployments in
environments which require a proxy server.
While Squid is a production-capable proxy server, this role does very
basic configuration and is not optimized for serving large volumes of
traffic.