Default to building Slurm with dynamic libs
Previously we built Slurm with statically-linked binaries, but this
turned out not to have the advantages we hoped. In addition, it turns
out this has caused problems with certain Slurm commands such as `sstat`
(see #720).

This commit swaps the configure flag `--without-shared-libslurm` to
`--with-shared-libslurm`. After testing, I can confirm the `sstat`
command seems to work again.