Add Molecule testing for Singularity, plus infra for more roles
Summary
-------

- Fixes the ability to install Singularity on EL8 by enabling the
  necessary repos
- Move the pre-execution logic for the Singularity Galaxy role into a
  new DeepOps role, singularity_wrapper
- Modify the playbook to use singularity_wrapper
- Add Molecule tests to singularity_wrapper role
- Move Galaxy requirements for singularity wrapper into the
  singularity_wrapper role, and include this from
  `roles/requirements.yml`
- Add a Github Action configuration to execute this and other Molecule
  tests of DeepOps roles
- Add documentation to `docs/deepops/testing.md` to describe these tests
  and provide instructions on adding more

Test plan
---------

Clean execution of the Github action for this role

Future work
-----------

This change adds a lot of infrastructure that should make it easier to
add testing to additional roles. Future changes should enable testing of
more of the roles in DeepOps, but I didn't want to get *too* carried
away in this one. :wink: