Break out functional host groups for slurm cluster
Summary
-------

- Add new hostgroups to the example inventory file based on the
different functional node types we expect on a Slurm cluster
- Use those hostgroup names as the defaults when selecting which hosts
to run on in the slurm-cluster.yml playbook

Detail
------

DeepOps supports breaking out Slurm cluster topology in a few different
ways.

This includes simple clusters with a single "head node" that runs all
the services like slurmctld, nfs, and monitoring... All the way up to
SuperPODs, where we have 5 different types of management node and split
up the functionality between them.

We support all of this in a single slurm-cluster.yml playbook by using
include tasks that look like this:

```
- include: generic/nfs-server.yml
  vars:
    hostlist: "{{ nfs_server_group | default('slurm-master[0]') }}"
  when: slurm_enable_nfs_server
```

And the playbook uses that hostlist var like so to decide what hosts to
run on:

```
---
- hosts: "{{ hostlist | default('all') }}"
  become: yes
  roles:
    - { role: nfs, nfs_is_server: yes }
  tags:
    - nfs_server
```
So that, for example, we default to putting the NFS server on the first
member of the slurm-master group.

But in theory we can define a variable nfs_server_group to be called
something like slurm-nfs, and put a different host in there.

Maksim Kunin discovered that this works when the variable is defined in
an --extra-vars file, but not when it's defined in group_vars. See issue
https://github.com/NVIDIA/deepops/issues/1085

I dug into this, and I think I know why.

When Ansible executes the included playbook, it appears to evaluate this
line:

```
- hosts: "{{ hostlist | default('all') }}"
```

to something like this:

```
- hosts: "{{ (slurm_monitoring_group | default('slurm-master[0]') ) | default('all') }}"
```

It turns out that Ansible only sources variables from group_vars/ when
it is executing on hosts that are members of the specified inventory
group.

However, when Ansible is evaluating the hosts: ... line of a play, it
doesn't know yet what hosts it will be running on. So it doesn't include
any of the variables defined in group_vars/ or host_vars/

Unfortunately, that means that slurm_monitoring_group is effectively
undefined here, even if it's defined in group_vars/slurm-cluster.yml.
Because Ansible hasn't decided it's running on a host in the
slurm-cluster group, it hasn't sourced that file.

--extra-vars works because it's specified at the level of the
ansible-playbook execution, so the variable is already defined.

This is really annoying! It's really helpful to be able to decide
dynamically which hosts to run on, and it would be nice to be able to
control this in our DeepOps config directly.

To solve this, we add new groups to the inventory based on our different
functional node types, and then use those group names in the playbook.