Temporarily disable CentOS 8 tests in Molecule
Starting on Jan 31, the CentOS 8 mirror lists are no longer working. This
is causing our CentOS 8 tests in Molecule to fail because we can't
install packages.

While I've tested using the CentOS Vault instead of the canonical mirror
list, this isn't a great solution:

- It's a bit on the slow side, because we don't have a global network of
mirrors
- CentOS 8 is no longer supported, so we'd just be targeting an
increasingly old distribution

We're discussing how to address this in our automated tests. But for
now, we'll disable these tests so that our other Molecule tests will
run.