Migrate CentOS-8 molecule tests to Rocky
Starting on Jan 31, the CentOS 8 mirror lists are no longer working.
This is causing our CentOS 8 tests in Molecule to fail because we can't
install packages.

While I've tested using the CentOS Vault instead of the canonical mirror
list, this isn't a great solution:

- It's a bit on the slow side, because we don't have a global network of mirrors
- CentOS 8 is no longer supported, so we'd just be targeting an increasingly old distribution

As a proposed solution, I've converted the molecule tests to use [Rocky
Linux](https://rockylinux.org/). This is mostly a drop-in replacement,
with some minor changes needed in the mofed, facts, and nhc roles due to the
Rocky container being a little more minimal than the CentOS one.