Remove hybrid cluster documentation
Hybrid clusters which run both Kubernetes and Slurm are not currently
well-supported or tested in DeepOps. NVIDIA's preferred solution for
this type of cluster is currently NVIDIA Bright Cluster Manager, which
provides a more robust solution to managing multiple workload managers
within the same cluster, and additionally supports UGE and PBS.

This PR makes the following changes:

- Clarify support in `README.md`
- Remove hybrid references in `docs/README.md`
- Completely remove `docs/deepops/dgx-pod.md`, which is also out of date
in several respects (such as specifying a Ceph storage system)

A good future project would be to rewrite the DGX POD doc based on an up
to date process with the latest DeepOps release.