Fix bugs preventing slurm reinstall or rebuild
Bugs:

1. docker.slurm-exporter.service creates subdirectories
`/usr/loca/bin/{sinfo,sdiag,squeue} when the excutables are absent
(due to uninstall process). When deepops tries to install the executables
again, they are put under the unwanted subdirectories (e.g.
/usr/local/bin/sinfo/sinfo instead of /usr/local/bin/sinfo)

2. Slurm 'make uninstall' doesn't remove files under "{{ slurm_install_prefix
}}/lib/slurm". Therefore the directory needs to be removed. A safer solution
would be to run the correct 'make uninstall' command (at the correct location)
to uninstall the files. Without this handling, slurm fails to start because of
library version conflicts.